refactor(capture-sdk): Merge OnboardingPageFragmentImpl with OnboardingPageFragmentCompat into OnboardingPageFragment
PIA-2290
