--- capture-sdk/sdk/src/main/java/net/gini/android/capture/onboarding/OnboardingPageFragmentCompat.java
@@ -1,73 +1,0 @@
-package net.gini.android.capture.onboarding;
-
-import android.app.Activity;
-import android.content.DialogInterface;
-import android.os.Bundle;
-import android.util.Log;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import net.gini.android.capture.internal.ui.FragmentImplCallback;
-import net.gini.android.capture.internal.util.AlertDialogHelperCompat;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-
-/**
- * Internal use only.
- *
- * @suppress
- */
-public class OnboardingPageFragmentCompat extends Fragment implements FragmentImplCallback {
-
-    private OnboardingPageFragmentImpl mFragmentImpl;
-
-    public static OnboardingPageFragmentCompat createInstance(@NonNull final OnboardingPage page, final boolean isLastPage) {
-        final OnboardingPageFragmentCompat fragment = new OnboardingPageFragmentCompat();
-        fragment.setArguments(OnboardingPageFragmentHelper.createArguments(page, isLastPage));
-        return fragment;
-    }
-
-    @Override
-    public void onCreate(final Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        mFragmentImpl = OnboardingPageFragmentHelper.createFragmentImpl(this, getArguments());
-    }
-
-    @Nullable
-    @Override
-    public View onCreateView(final LayoutInflater inflater, final ViewGroup container,
-            final Bundle savedInstanceState) {
-        return mFragmentImpl.onCreateView(inflater, container, savedInstanceState);
-    }
-
-    @Override
-    public void showAlertDialog(@NonNull final String message,
-            @NonNull final String positiveButtonTitle,
-            @NonNull final DialogInterface.OnClickListener positiveButtonClickListener,
-            @Nullable final String negativeButtonTitle,
-            @Nullable final DialogInterface.OnClickListener negativeButtonClickListener,
-            @Nullable final DialogInterface.OnCancelListener cancelListener) {
-        final Activity activity = getActivity();
-        if (activity == null) {
-            return;
-        }
-        AlertDialogHelperCompat.showAlertDialog(activity, message, positiveButtonTitle,
-                positiveButtonClickListener, negativeButtonTitle, negativeButtonClickListener,
-                cancelListener);
-    }
-
-    @Override
-    public void onResume() {
-        super.onResume();
-        mFragmentImpl.onResume();
-    }
-
-    @Override
-    public void onPause() {
-        super.onPause();
-        mFragmentImpl.onPause();
-    }
-}
